show zoom/reset buttons after brush
(already shown after mousewheel-zoom).
factor showAxisZoomResetButtons()  from brushHelper() in draw-map.js to utils/draw/axisBrush.js, and also call it in zoom().  rename vars zoomSwitch{S,E,} to g{S,E,}.
These moves are to enable showAxisZoomResetButtons() to access these library functions :
move axisFeatureCircles_selectAll() from  draw-map.js to utils/draw/axis.js
move I from draw-map.js  to d3-svg.js

 80a4c8a 239750 Sep 24 19:14  components/draw-map.js
 4e619b8  10543 Sep 24 16:48  utils/draw/axis.js
 192c6d1   3467 Sep 24 17:03  utils/draw/axisBrush.js
 a4803fc   1811 Sep 24 16:12  utils/draw/d3-svg.js
